python_server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlclassifier-server"
version = "0.1.0"
description = "Deep learning pixel classification server for QuPath"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "UW-LOCI", email = "loci@wisc.edu"}
]
keywords = ["deep-learning", "segmentation", "qupath", "microscopy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",  # Required for FastAPI form data handling
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "segmentation-models-pytorch>=0.3.3",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",  # Base package, works cross-platform
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "pydantic>=2.0.0",
    "tifffile>=2023.7.0",
    "imagecodecs>=2023.7.0",
    "albumentations>=1.3.0",
]

[project.optional-dependencies]
cuda = [
    "onnxruntime-gpu>=1.16.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "scipy>=1.10.0",  # For test data generation
]

[project.scripts]
dlclassifier-server = "dlclassifier_server.main:run"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
